server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
jinja2>=3.1.2
python-multipart>=0.0.6
email-validator>=2.0.0
numpy>=1.24.0
apscheduler>=3.10.0